+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.controller.RaceController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Main main = new Main();
-        main.run();
+        RaceController raceController = new RaceController();
+        raceController.run();
     }
 }
