+++ src/main/java/racingcar/controller/CarController.java
@@ -1,11 +1,18 @@
 package racingcar.controller;
 
-import racingcar.Car;
+import racingcar.model.Car;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 public class CarController {
+
+    public List<String> splitCarNames(String input) {
+        List<String> carNames = Arrays.stream(input.split(",")).toList();
+        return carNames;
+    }
+
     public Car createSingleCar(String carName) {
         Car car = new Car(carName, 0);
         return car;
